feat: Make AudioPlayerFragment completly Make AudioPlayerFragment completly and adjust these fragment to MainActivity
